requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-compat==1.0.15
django-cors-headers==3.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
numpy==1.18.2
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
xlwt==1.3.0
gunicorn==20.0.4
